feat(parser): move definitions to their own modules
