+++ crates/aiken-lang/src/parser/definitions/mod.rs
@@ -1,0 +1,15 @@
+mod constant;
+mod data_type;
+mod function;
+mod import;
+mod test;
+mod type_alias;
+mod validator;
+
+pub use constant::parser as constant;
+pub use data_type::parser as data_type;
+pub use function::parser as function;
+pub use import::parser as import;
+pub use test::parser as test;
+pub use type_alias::parser as type_alias;
+pub use validator::parser as validator;
