+++ crates/aiken-lang/src/parser/definitions/validator.rs
@@ -1,0 +1,58 @@
+use chumsky::prelude::*;
+
+use crate::{
+    ast,
+    parser::{error::ParseError, token::Token},
+};
+
+pub fn parser() -> impl Parser<Token, ast::UntypedDefinition, Error = ParseError> {
+    just(Token::Validator)
+        .ignore_then(
+            fn_param_parser(true)
+                .separated_by(just(Token::Comma))
+                .allow_trailing()
+                .delimited_by(just(Token::LeftParen), just(Token::RightParen))
+                .map_with_span(|arguments, span| (arguments, span))
+                .or_not(),
+        )
+        .then(
+            super::function()
+                .repeated()
+                .at_least(1)
+                .at_most(2)
+                .delimited_by(just(Token::LeftBrace), just(Token::RightBrace))
+                .map(|defs| {
+                    defs.into_iter().map(|def| {
+                        let ast::UntypedDefinition::Fn(fun) = def else {
+                            unreachable!("It should be a fn definition");
+                        };
+
+                        fun
+                    })
+                }),
+        )
+        .map_with_span(|(opt_extra_params, mut functions), span| {
+            let (params, params_span) = opt_extra_params.unwrap_or((
+                vec![],
+                ast::Span {
+                    start: 0,
+                    end: span.start + "validator".len(),
+                },
+            ));
+
+            ast::UntypedDefinition::Validator(ast::Validator {
+                doc: None,
+                fun: functions
+                    .next()
+                    .expect("unwrapping safe because there's 'at_least(1)' function"),
+                other_fun: functions.next(),
+                location: ast::Span {
+                    start: span.start,
+                    // capture the span from the optional params
+                    end: params_span.end,
+                },
+                params,
+                end_position: span.end - 1,
+            })
+        })
+}
